We longer cache inner args. This fixes the no result crash bug
Note that this does not effect the base coverage of the ltp starting
corpus (~16k). I could not determine the exact cause of the bug, and
after one run this change does not seem to effect the overall coverage
via fuzzing.

From looking at crash logs, every crash seemed to result from a program
being mutated such that a parent arg <rN = ...> was converted into a
normal ConstArg, i.e. it's r.Uses map was flushed. Hence, when we
reached its child rN, the dependency was not set up and the program
could not be serialized. I hypothesize the cause to be somewhere in
mutation.go

A dirty fix would be to validate the program every time before
serializing, and reconstruct these dependencies. possible TODO?
But again, removing the caching doesn't seem to effect performance.